Address review: move accessor out of coremod, keep conversion behaviour
- LivingEntityGetDeathSoundAccessor does not replace a transformer, so it
  does not belong in the asm.mixin.coremod package; move it to the asm.mixin
  root
- convertEntity: restore the vanilla finalizeSpawn call (the NeoForge
  finalizeMobSpawn event was dropped; this keeps the conversion
  initialisation behaviour)
